Bugs are fixed and encryption is in progress

+++ App/Backend/Modules/Author/AuthorController.php
@@ -14,24 +14,35 @@ class AuthorController extends BackController
   public function executeIndex(HTTPRequest $request)
   {
     $this->page->addVar('title', 'Gestion des auteurs');
-
     $manager = $this->managers->getManagerOf('Authors');
-
     $this->page->addVar('listeAuthors', $manager->getList());
     $this->page->addVar('nombreAuthor', $manager->count());
   }
 
   public function executeInsert(HTTPRequest $request)
   {
-    $this->processForm($request);
-    $this->page->addVar('title', 'Ajout d\'un auteur');
+    if(!$this->app->user()->isAuthenticatedAdmin(true))
+    {
+      $this->app->httpResponse()->redirect('/admin/');
+    }
+    else
+    {
+      $this->processForm($request);
+      $this->page->addVar('title', 'Ajout d\'un auteur');
+    }
   }
 
   public function executeUpdate(HTTPRequest $request)
   {
-    $this->processForm($request);
- 
-    $this->page->addVar('title', 'Modification d\'un auteur');
+    if(!$this->app->user()->isAuthenticatedAdmin(true))
+    {
+      $this->app->httpResponse()->redirect('/admin/');
+    }
+    else
+    {
+      $this->processForm($request); 
+      $this->page->addVar('title', 'Modification d\'un auteur');
+    }
   }
   
   public function processForm(HTTPRequest $request)
@@ -40,13 +51,23 @@ class AuthorController extends BackController
 
     if ($request->method() == 'POST')
     {
+      $salt = "$2y$14$";
+      for ($i = 0; $i < 22; $i++) 
+      {
+        $salt .= substr("./ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789", mt_rand(0, 63), 1);
+      }
+      // crypt($request->postData('password'), $author->salt);
+      // crypt($request->postData('passwordCheck'), $author->salt);
       $author = new Author([
-        'firstname' => $request->postData('firstname'),
-        'lastname' => $request->postData('lastname'),
+        'firstname'   => $request->postData('firstname'),
+        'lastname'    => $request->postData('lastname'),
         'dateofbirth' => $request->postData('dateofbirth'),
-        'pseudo' => $request->postData('pseudo'),
-        'pwd' => $request->postData('pwd'),
-        'type' => $request->postData('type'),
+        'pseudo'      => $request->postData('pseudo'),
+        'pwd'         => $request->postData('password'),
+        'pwdCheck'    => $request->postData('passwordCheck'),
+        'type'        => $request->postData('type'),
+        'email'       => $request->postData('email'),
+        'salt'        => $salt,
       ]);
       if ($request->getExists('id'))
       {
@@ -68,7 +89,7 @@ class AuthorController extends BackController
     
     $formBuilder = new AuthorFormBuilder($author);
 
-    $formBuilder->build();
+    $formBuilder->build(/*Passer les id de types*/);
 
     $form = $formBuilder->form();
 
@@ -90,14 +111,16 @@ class AuthorController extends BackController
     if ($request->method() == 'POST')
     {
       $news = new Author([
-        'id' => $request->getData('id'),
-        'firstname' => $request->postData('firstname'),
-        'lastname' => $request->postData('lastname'),
-        'registrationdate' => $request->postData('registrationdate'),
-        'dateofbirth' => $request->postData('dateofbirth'),
-        'dateModif' => $request->postData('dateModif'),
-        'pwd' => $request->postData('pwd'),
-        'pseudo' => $request->postData('pseudo'),
+        'BAC_id' => $request->getData('BAC_id'),
+        'BAC_firstname' => $request->postData('BAC_firstname'),
+        'BAC_lastname' => $request->postData('BAC_lastname'),
+        'BAC_registrationdate' => $request->postData('BAC_registrationdate'),
+        'BAC_dateofbirth' => $request->postData('BAC_dateofbirth'),
+        'BAC_dateModif' => $request->postData('BAC_dateModif'),
+        'BAC_password' => $request->postData('BAC_password'),
+        'BAC_passwordCheck ' => $request->postData('BAC_passwordCheck'),
+        'BAC_pseudo' => $request->postData('BAC_pseudo'),
+        'BAC_email' => $request->postData('BAC_email'),
       ]);
     }
     else
@@ -122,12 +145,17 @@ class AuthorController extends BackController
 
   public function executeDelete(HTTPRequest $request)
   {
-    $authorId = $request->getData('id');
-    
-    $this->managers->getManagerOf('Authors')->delete($authorId);
-
-    $this->app->user()->setFlash('L\'auteur a bien été supprimée !');
-    $this->app->httpResponse()->redirect('.');
+    if(!$this->app->user()->isAuthenticatedAdmin(true))
+    {
+      $this->app->httpResponse()->redirect('/admin/');
+    }
+    else
+    {
+      $authorId = $request->getData('id');
+      $this->managers->getManagerOf('Authors')->delete($authorId);
+      $this->app->user()->setFlash('L\'auteur a bien été supprimée !');
+      $this->app->httpResponse()->redirect('/admin/author-show.html');
+    }
   }
 
 
@@ -142,6 +170,26 @@ public function executeShow(HTTPRequest $request)
     }
     $this->page->addVar('authors', $authors);
     // $this->page->addVar('comments', $this->managers->getManagerOf('Comments')->getListOf($news->id()));
+  }
+
+  //List of all authors registered in the database
+public function executePasswordrecovering(HTTPRequest $request)
+  {
+    $login = $request->postData('login');
+    $email = $request->postData('email');
+
+    $author = $this->managers->getManagerOf('Authors')->getUnique($login); 
+    if(is_null($author) || $author->email() != $email)
+    {
+      $message = "Le login ou l'email est incorrect";
+    }
+    else
+    {
+      $message = "Un email contenant votre mot de passe vous a été envoyé";
+      // envoi du mail
+      mail($author['email'], 'Récupération du mot de passe', $author->email());
+      $this->page->addVar('message', $message);
+    }
   }
 }
